refactor(cli): extract run-tests/run-build into commands/ops_commands.py
Moves the subprocess-wrapping ritual commands out of cli.py into a
dedicated module. Both still thread mobile_task_notification and
refresh the tmux HUD via update_tmux_mobile_indicator — imports
are duplicated from cli.py, not cross-referenced.

`run-tests --help` and `run-build --help` both render cleanly via
CliRunner. CLI test suite shows the same 4 pre-existing failures; no
new regressions.
